requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.1
python-docx>=1.0.0
openai>=1.0.0
requests>=2.31.0
